Quick test addition and renames
