+++ src/k_startup/Makefile
@@ -18,7 +18,7 @@ _SRCS 		?= kernel.c \
 			   page.c \
 			   page_helpers.c \
 			   ata_block_device.c \
-			   vga_term.c
+			   vga_cd.c
 
 # REQUIRED: Names (NOT PATHS) of all .s files found in the src folder
 _ASMS		?= boot.s \
