+++ src/k_startup/src/plugin_kb.c
@@ -25,7 +25,7 @@ plugin_t *new_plugin_kb(kernel_state_t *ks) {
     return (plugin_t *)plg_kb;
 }
 
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 
 static fernos_error_t plg_kb_kernel_cmd(plugin_t *plg, plugin_kernel_cmd_id_t kcmd, uint32_t arg0,
         uint32_t arg1, uint32_t arg2, uint32_t arg3) {
