+++ src/k_startup/test/process.c
@@ -6,7 +6,7 @@
 
 #include "k_startup/page.h"
 
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 
 static bool pretest(void);
 static bool posttest(void);
