+++ src/s_block_device/test/fat32.c
@@ -9,7 +9,7 @@
 #include "s_block_device/block_device.h"
 #include "s_block_device/mem_block_device.h"
 #include "s_block_device/fat32.h"
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 
 static bool pretest(void);
 static bool posttest(void);
