+++ src/s_block_device/test/file_sys.c
@@ -1,7 +1,7 @@
 
 #include "s_block_device/test/file_sys.h"
 
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 #include "s_mem/allocator.h"
 #include "s_util/str.h"
 #include "s_util/rand.h"
