+++ src/s_mem/src/simple_heap.c
@@ -1,7 +1,7 @@
 
 
 #include "s_mem/simple_heap.h"
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 #include "s_mem/allocator.h"
 #include "s_util/err.h"
 #include "s_util/misc.h"
