+++ src/s_mem/test/allocator.c
@@ -3,7 +3,7 @@
 #include "s_util/str.h"
 #include "s_mem/test/allocator.h"
 #include "s_mem/allocator.h"
-#include "k_startup/vga_term.h"
+#include "k_startup/vga_cd.h"
 
 
 static bool pretest(void);
